+++ internal/logdb/kv/rocksdb/gorocksdb/dynflag.go
@@ -1,4 +1,0 @@
-package gorocksdb
-
-// #cgo LDFLAGS: -lrocksdb -lstdc++ -lm
-import "C"