+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Dragonboat v3.4 comes with many improvements. All v3.3.x users are recommended t
 - Fixed unreachable notification.
 - Upgraded to a more recent version of pebble.
 - Non-voting node (used to be called observer node) support has been marked as production ready.
+- Made the experimental gossip feature a first class citizen of the library.
 
 ### Other changes
 
@@ -31,7 +32,7 @@ Dragonboat v3.3 is a major release that comes with new features and improvements
 - Pebble, which is bidirectional compatible with RocksDB, has been made the default engine for storing Raft Logs. RocksDB and CGO are no longer required.
 - Added the ability to slow down incoming proposals when the Raft Logs engine is highly loaded.
 - Added the option to get notified when proposals and config changes are committed.
-- Added a gossip service to allow NodeHosts to use dynamically assigned IP addresses as RaftAddress.
+- Added an experimental gossip service to allow NodeHosts to use dynamically assigned IP addresses as RaftAddress.
 - Added the ability to better control memory footprint.
 - Added ARM64/Linux as a new targeted platform.